refactor(admin/customer): edit form (#176)
