+++ client/src/admin/customer/customer-edit.jsx
@@ -1,91 +1,127 @@
+import { useCallback } from 'react';
 import {
   Edit,
   SimpleForm,
   TextInput,
-  SelectInput,
   required,
-  ReferenceInput,
   DateInput,
   NumberInput,
+  useNotify,
+  useUpdate,
+  useRedirect
 } from "react-admin";
 import { Box } from "@mui/material";
 import { CustomToolbar } from "../custom-toolbar";
 
-export const CustomerEdit = () => (
-  <Edit>
-    <SimpleForm toolbar={<CustomToolbar />}>
-      <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
-        <Box flex={2} mr={{ xs: 0, sm: "0.5em" }}>
-          <TextInput source="name" label="Nombre" validate={required()} fullWidth />
+export const CustomerEdit = () => {
+  const redirect = useRedirect();
+  const notify = useNotify();
+
+  const [update] = useUpdate();
+  const save = useCallback(
+      async values => {
+          try {
+
+              console.log('New data: ');
+              console.table(values);
+              await update(
+                  'customers',
+                  { id: values.id, data: values },
+                  { returnPromise: true }
+              );
+
+              notify('ra.notification.updated', {
+                  type: 'info',
+                  messageArgs: { smart_count: 1 },
+              });
+              redirect('list', 'customers');
+          } catch (error) {
+              console.table(error);
+              if (error.body && error.body.errors) {
+                  return error.body.errors;
+              }
+
+              if (error.body && error.body.validationError) {
+                  notify(error.body.validationError, {
+                      type: 'error'
+                  });
+                  return '';
+              } else {
+                  notify('Ooops! Error validating data', {
+                      type: 'error'
+                  });
+                  return '';
+              }
+          }
+      },
+      [update, notify, redirect]
+  );
+
+  return (
+    <Edit title="Cuenta">
+      <SimpleForm toolbar={<CustomToolbar />} onSubmit={save}>
+        <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
+          <Box flex={2}>
+            <TextInput
+              source="name"
+              label="Nombre y Apellido"
+              validate={required()}
+              fullWidth
+            />
+          </Box>
+          <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
+            <DateInput source="birthDate" label="Fecha de nacimiento" fullWidth />
+          </Box>
         </Box>
-        <Box flex={2} ml={{ xs: 0, sm: "0.5em" }}>
-          <ReferenceInput source="userId" label="Usuario" reference="users">
-            <SelectInput validate={required()} fullWidth />
-          </ReferenceInput>
+        <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
+          <Box flex={1}>
+            <TextInput
+              source="address"
+              label="Dirección"
+              fullWidth
+            />
+          </Box>
+          <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
+            <TextInput
+              source="city"
+              label="Ciudad"
+              fullWidth
+            />
+          </Box>
         </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <SelectInput
-            source="status"
-            label="Estado"
-            optionValue="name"
-            defaultValue={"Active"}
-            choices={[
-              { id: "Active", name: "Active" },
-              { id: "Disabled", name: "Disabled" },
-            ]}
-            validate={required()}
-            fullWidth
-          />
+        <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
+          <Box flex={2}>
+            <TextInput
+              source="state"
+              label="Provincia"
+              fullWidth
+            />
+          </Box>
+          <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
+            <NumberInput
+              source="zip"
+              label="Código postal"
+              fullWidth
+            />
+          </Box>
         </Box>
-      </Box>
-      <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <TextInput
-            source="address"
-            label="Dirección"
-            validate={required()}
-            fullWidth
-          />
+        <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
+          <Box flex={1}>
+            <TextInput
+              source="telephone"
+              label="Teléfono"
+              validate={required()}
+              fullWidth />
+          </Box>
+          <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
+            <NumberInput
+              source="document"
+              label="Documento"
+              validate={required()}
+              fullWidth />
+          </Box>
         </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <TextInput
-            source="city"
-            label="Ciudad"
-            validate={required()}
-            fullWidth
-          />
-        </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <TextInput
-            source="state"
-            label="Provincia"
-            validate={required()}
-            fullWidth
-          />
-        </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <NumberInput
-            source="zip"
-            label="Código postal"
-            validate={required()}
-            fullWidth
-          />
-        </Box>
-      </Box>
-      <Box display={{ xs: "block", sm: "flex", width: "100%" }}>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <TextInput source="email" label="Email" validate={required()} fullWidth />
-        </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <TextInput source="telephone" label="Teléfono" fullWidth />
-        </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <NumberInput source="document" label="Documento" fullWidth />
-        </Box>
-        <Box flex={1} ml={{ xs: 0, sm: "0.5em" }}>
-          <DateInput source="birthDate" label="Fecha de nacimiento" fullWidth />
-        </Box>
-      </Box>
-    </SimpleForm>
-  </Edit>
-);
+      </SimpleForm>
+    </Edit>
+  );
+}
