+++ client/src/admin/customer/customer-list.jsx
@@ -3,12 +3,10 @@ import {
   List,
   TextField,
   TextInput,
-  SelectInput,
-  ReferenceField,
   DateField,
   NumberField,
-  EmailField,
   SimpleList,
+  ReferenceField
 } from "react-admin";
 import { useMediaQuery } from "@mui/material";
 
@@ -16,16 +14,7 @@ export const CustomerList = () => {
   const isSmall = useMediaQuery((theme) => theme.breakpoints.down("sm"));
 
   const customerFilters = [
-    <TextInput source="name" label="Buscar" alwaysOn />,
-    <SelectInput
-      source="status"
-      label="Estado"
-      choices={[
-        { id: "Active", name: "Active" },
-        { id: "Disabled", name: "Disabled" },
-      ]}
-      alwaysOn
-    />,
+    <TextInput source="name" alwaysOn />,
   ];
   return (
     <List filters={customerFilters}>
@@ -37,19 +26,17 @@ export const CustomerList = () => {
         />
       ) : (
         <Datagrid rowClick="edit" bulkActionButtons={false}>
-          <TextField source="name" label="Nombre"/>
-          <ReferenceField source="userId" label="Usuario" reference="users">
+          <TextField source="name" />
+          <TextField source="address" />
+          <TextField source="city" />
+          <TextField source="state" />
+          <NumberField source="zip" />
+          <ReferenceField source="userId" reference="users">
             <TextField source="userName" />
-          </ReferenceField>
-          <TextField source="address" label="Dirección" />
-          <TextField source="city" label="Ciudad" />
-          <TextField source="state" label="Provincia" />
-          <NumberField source="zip" label="Código Postal" />
-          <EmailField source="email" label="Email" />
-          <TextField source="telephone" label="Teléfono" />
-          <TextField source="document" label="Documento" />
-          <DateField source="birthDate" label="Fecha de nacimiento" />
-          <TextField source="status" label="Estado" />
+          </ ReferenceField>
+          <TextField source="telephone" />
+          <TextField source="document" />
+          <DateField source="birthDate" />
         </Datagrid>
       )}
     </List>
