+++ client/src/miCuenta/index.js
@@ -2,7 +2,7 @@ import { Admin, Resource } from 'react-admin';
 import { dataProvider } from './data-provider';
 import { YazzLayout } from './layout';
 import { CustomerList } from './customer/customer-list';
-import { CustomerEdit } from './customer/customer-edit';
+import { CustomerEdit } from '../admin/customer/customer-edit';
 import { ReviewList } from './review/review-list';
 import { ReviewShow } from './review/review-show';
 import { OrderList } from './order/order-list';
